isolate working directory changing test case
refactor a ServeDist test case that changes CWD during testing into a
standalone test program to prevent it from polluting CWD of other test
cases.

Fix #73
